Rewrite systemd integration as a plugin
The primary motiviation for this is that plugins have native integration
for background threads. This is much cleaner since it allows tracking of
those. For example, it's possible to clean up those threads in the test
suite.
